fix: trim long strings when printing dataframe
