+++ tests/repr_test.py
@@ -23,3 +23,11 @@ def test_mask(ds_local):
 def test_repr_expression(ds_local):
     df = ds_local
     assert 'Error' not in repr(df.x)
+
+def test_repr_df_long_string():
+    long_string = "Hi there" * 100
+    df = vaex.from_arrays(s=[long_string] * 100)
+    assert long_string not in repr(df)
+    assert long_string not in str(df)
+    assert long_string not in df._repr_html_()
+    assert long_string not in df._as_html_table(0, 10)
